airflow_lab/requirements.txt
pandas
pyarrow
sqlalchemy
psycopg2-binary
apache-airflow-providers-postgres